Create Basic_Speed_2_Simulator
This code simulates a speed 2 robber trying to avoid a specified number of cops on a grid of a specified size.  The cops move randomly, except when able to capture the robber.  The robber tries to avoid the cops and avoid edges and corners (based on a local potential function).